requirements.txt
numpy
pandas
scikit-learn
tensorflow
streamlit
paho-mqtt
plotly